Refactor: 변경, findUser 위치
